// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/outsourcedDataTaskStandardBasicDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程
 * OpenAPI spec version: v1.0.0
 */
import type { OutsourcedDataTaskStandardBasicRowDTO } from "./outsourcedDataTaskStandardBasicRowDTO";
import type { OutsourcedDataTaskStandardRawColumnDTO } from "./outsourcedDataTaskStandardRawColumnDTO";

export interface OutsourcedDataTaskStandardBasicDTO {
  /** No comments found. */
  basicInfoCount?: number;
  /** No comments found. */
  basicRows?: OutsourcedDataTaskStandardBasicRowDTO[];
  /** No comments found. */
  batchId?: string;
  /** No comments found. */
  dataStartRowNumber?: number;
  /** No comments found. */
  fileId?: number;
  /** No comments found. */
  headerRowNumber?: number;
  /** No comments found. */
  metricCount?: number;
  /** No comments found. */
  rawColumns?: OutsourcedDataTaskStandardRawColumnDTO[];
  /** No comments found. */
  sheetName?: string;
  /** No comments found. */
  subjectCount?: number;
  /** No comments found. */
  taskId?: number;
  /** No comments found. */
  title?: string;
  /** No comments found. */
  valuationId?: number;
  /** No comments found. */
  workbookPath?: string;
}